moved robbing and trading into player card moved player card into playercards
