+++ src/components/TimerAlert.js
@@ -4,7 +4,7 @@ import api from '../lib/api'
 import Auth from '../lib/auth'
 import { socket } from '../lib/sockets'
 
-const TimerAlert = ({gameEvents}) => {
+const TimerAlert = ({ gameEvents }) => {
   const [createdBy, setCreatedBy] = useState(null)
 
   useEffect(() => {
