+++ src/lib/auth.js
@@ -1,20 +1,20 @@
 class Auth {
 
-  static setToken(token) {
-    localStorage.setItem('token', token)
-  }
+	static setToken(token) {
+		localStorage.setItem('token', token)
+	}
 
-  static getToken() {
-    return localStorage.getItem('token')
-  }
+	static getToken() {
+		return localStorage.getItem('token')
+	}
 
-  static isAuthenticated() {
-    return !!this.getToken()
-  }
+	static isAuthenticated() {
+		return !!this.getToken()
+	}
 
-  static logout() {
-    localStorage.removeItem('token')
-  }
+	static logout() {
+		localStorage.removeItem('token')
+	}
 
 }
